# openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "BioCLIP Local Server",
		"version": "0.1.0"
	},
	"paths": {
		"/api/maintenance/run": {
			"post": {
				"summary": "Trigger Maintenance",
				"operationId": "trigger_maintenance_api_maintenance_run_post",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				}
			}
		},
		"/": {
			"get": {
				"summary": "Read Index",
				"operationId": "read_index__get",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"text/html": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				}
			}
		},
		"/api/identify": {
			"post": {
				"summary": "Identify Image",
				"operationId": "identify_image_api_identify_post",
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"schema": {
								"$ref": "#/components/schemas/Body_identify_image_api_identify_post"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/api/identify/online": {
			"post": {
				"summary": "Identify Image Online",
				"operationId": "identify_image_online_api_identify_online_post",
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"schema": {
								"$ref": "#/components/schemas/Body_identify_image_online_api_identify_online_post"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"Body_identify_image_api_identify_post": {
				"properties": {
					"image": {
						"type": "string",
						"contentMediaType": "application/octet-stream",
						"title": "Image"
					},
					"province": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Province"
					},
					"lat": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Lat"
					},
					"lng": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Lng"
					}
				},
				"type": "object",
				"required": [
					"image"
				],
				"title": "Body_identify_image_api_identify_post"
			},
			"Body_identify_image_online_api_identify_online_post": {
				"properties": {
					"image": {
						"type": "string",
						"contentMediaType": "application/octet-stream",
						"title": "Image"
					},
					"token": {
						"type": "string",
						"title": "Token"
					},
					"province": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Province"
					},
					"lat": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Lat"
					},
					"lng": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "null"
							}
						],
						"title": "Lng"
					}
				},
				"type": "object",
				"required": [
					"image",
					"token"
				],
				"title": "Body_identify_image_online_api_identify_online_post"
			},
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": {
							"$ref": "#/components/schemas/ValidationError"
						},
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
			"ValidationError": {
				"properties": {
					"loc": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "integer"
								}
							]
						},
						"type": "array",
						"title": "Location"
					},
					"msg": {
						"type": "string",
						"title": "Message"
					},
					"type": {
						"type": "string",
						"title": "Error Type"
					},
					"input": {
						"title": "Input"
					},
					"ctx": {
						"type": "object",
						"title": "Context"
					}
				},
				"type": "object",
				"required": [
					"loc",
					"msg",
					"type"
				],
				"title": "ValidationError"
			}
		}
	}
}